2-int_index.c

+++ 0x0F-function_pointers/2-int_index.c
@@ -10,7 +10,7 @@
  */
 int int_index(int *array, int size, int (*cmp)(int))
 {
-	size_t y;
+	int y;
 
 	if (array == NULL || cmp == NULL)
 		return (-1);
